requirements.txt
flask
requests
openai
gspread
oauth2client
python-dotenv
gunicorn
